testinggg/requirements.txt
accelerate==0.0.1
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
MouseInfo==0.1.3
nest-asyncio==1.6.0
numpy==2.2.2
openai==1.60.2
packaging==24.2
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytweening==1.2.0
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.1
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
